# docs/openapi/iamnande/cardmod/card/v1/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "iamnande/cardmod/card/v1/api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "CardAPI"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/cards": {
      "get": {
        "summary": "ListCards lists all available card resources.",
        "operationId": "CardAPI_ListCards",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListCardsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "filter.query",
            "description": "query is the actual search query.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "CardAPI"
        ]
      }
    },
    "/v1/cards/{cardId}": {
      "get": {
        "summary": "GetCard describes a single card.",
        "operationId": "CardAPI_GetCard",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetCardResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "cardId",
            "description": "card_id is the ID of the card to describe.",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "CardAPI"
        ]
      }
    }
  },
  "definitions": {
    "ListCardsRequestFilter": {
      "type": "object",
      "properties": {
        "query": {
          "type": "string",
          "description": "query is the actual search query."
        }
      },
      "description": "Filter is an optional search query to limit results with."
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Card": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "ID of the card"
        },
        "name": {
          "type": "string",
          "title": "Name of the card"
        }
      },
      "title": "A card resource"
    },
    "v1GetCardResponse": {
      "type": "object",
      "properties": {
        "card": {
          "$ref": "#/definitions/v1Card",
          "description": "card is the card entity that was described."
        }
      },
      "description": "GetCardResponse is the response model for GetCard."
    },
    "v1ListCardsResponse": {
      "type": "object",
      "properties": {
        "cards": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Card"
          },
          "description": "cards is the list of cards."
        }
      },
      "description": "ListCardsResponse is the response model for ListCards."
    }
  }
}